Fix #274: Avoid possible deadlock of timer callback
If the tick_time from the wait routine was perpetually
greater than the interval time, eventually the wait_time
became such that it was always below zero.  Once this
occurs, application callbacks would cease entirely.

To avoid this, run the callback condition in a loop
for periodic timers only (not for one-shot config).