Fix #287: cppcheck warnings
Replace use of (1<<31) and (INT_MAX) with (UINT32_MAX/2) which should
squelch the cppcheck warnings about the left shift as well as being
more correct (the values in question are uint32's, not int's).

Also address two false positives about a return value being unassigned,
by simply initializing the value to be 0.