Fix #297: Refactor OSAL C library stubs
Rather than utilizing a separate "stub-map-to-real.h" file that would
need to be injected into the build of the unit under test, this builds
the mappings into the mapping into the override files themselves.

The stub functions are then moved into a separate header under a
unique name with an OCS_ prefix.

This introduces a number of new files, as what previously required
a single header now requires two, but it it eliminates the separate
map file.