Fix #337: fix memory corruption produced by misplaced memset()
Calling memset too late after assigning the `sem` variable in the code of
`OS_BinSemCreate_Impl` caused the data pointed to by `sem` to get corrupted.

The issue was not caught by the existing test suite for POSIX OSAL when running
on Linux. However running the test suite on macOS revealed the anomalies
in the behavior of `pthread_cond_destroy()` which was working on the corrupted
memory as was demonstrated in:

Calling pthread_cond_destroy results in “Function not implemented” ENOSYS on macOS,
https://stackoverflow.com/questions/59560940/calling-pthread-cond-destroy-results-in-function-not-implemented-enosys-on-mac?noredirect=1#comment105301077_59560940

The original commit that introduced the issue is:

https://github.com/nasa/osal/commit/bfa7a33020e9a77221352a99bd2dbee4860ba852