Fix #135, Add separate CMakeLists.txt for each implementation
Do not use aux_source_directory to assemble a list of source files.
Instead, put a proper CMakeLists.txt file in each implementation
and build the implementation separately from the shared/common parts.

In addition to avoiding the aux_source_directory this allows PSP-specific
compile definitions to be set on a per-implementation basis because
it is defined separately.