Fix #432, Infer OSAL_SYSTEM_OSTYPE from OSAL_SYSTEM_BSPTYPE
The OSAL_SYSTEM_BSPTYPE specification by itself is enough to
build OSAL, as it implies a specific OS layer.

However, if OSAL_SYSTEM_OSTYPE is explicitly specified, it is used,
but will generate a warning if it is different than the OS layer
implied by the BSP, as there could be reasons for doing this
during development.