Fix #235, add check against OS_MAX_QUEUE_DEPTH
The OS_QueueCreate() function will now sanity check the depth
parameter against the configured OS_MAX_QUEUE_DEPTH value.  If
it is too large, an error will be returned.  This is a hard limit
and independent of the "permissive" mode.

The OS_MAX_QUEUE_DEPTH should be configured to the largest value
that an application may reasonably request.