Fix #456, improve filesystem type identification
Add a distinct type for FS_BASED entries, and an UNKNOWN value
for when a type is actually not yet identified.

This can be used to identify the FS type first at the shared
layer, then in the implementation layer as a fallback option
if not identifiable in the shared layer.

Use the volume name prefix "RAM" as a hint that the volume
is supposed to be a VOLATILE disk as opposed to a normal disk.
This is done in shared layer so it applies to all OS types.