Fix #450, add external source directory for OS/BSP
If the "OSAL_EXT_SOURCE_DIR" cache variable is set, this location
will be checked first for a BSP/OS implementation layer.  This can
point to an out-of-tree implementation layer if necessary.

However it is discouraged from actually doing this as there is no
attempt at API stability at the low level implementation layer.