Fix #479, add name association to stub arguments
Add the capability to store parameter names in addtion
to pointers as part of the stub argument context data.

Macroize the UT_Stub_RegisterContext function utstubs.h to
also pass the parameter name in addition to the pointer.

Also add hook-side accessor functions/macros to make it
easier and more reliable to get this information from
the context data.