Fix #501, add PIC library for ut assert
Add a position independent code (PIC) variant of the ut_assert library,
which can be dynamically loaded into other applications rather than running
as a standalone OSAL application.  This enables loading UT assert as a
CFE library.

This required moving some symbols/functions around.  Mainly, the PIC
library does _not_ contain the "utbsp.c" file which has the bindings
to the OSAL BSP when running as a standalone app, but everything else
should be the same.