Fix #506, enforce nonzero stack size
Resolve inconsistency in how the stack size is treated across
different OS implemntations.  POSIX would enforce a minimum,
where RTEMS would not.

With this change the user requested size is passed through
to the underlying OS exactly as is, no enforced minimum.

An additional sanity check is added at the shared layer to
ensure that the stack size is never passed as 0.